requirements.txt
pandas>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
openai>=1.0.0
python-dotenv>=1.0.0
wandb>=0.16.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-classic>=0.3.0
langchain-core>=0.3.0
langchain-core-documents>=0.3.0
langchain-core-prompts>=0.3.0
langchain-core-runnables>=0.3.0
langchain-core-runnables-prompts>=0.3.0

